pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "desktop-agent"
version = "2.1.0"
description = "Desktop automation CLI for AI agents with task recording and semantic search"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Mal", email = "excido@gmail.com"}
]
keywords = ["automation", "desktop", "ai", "cli", "atspi", "task-recording"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pillow>=9.0.0",
    "pytesseract>=0.3.10",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
]

[project.scripts]
desktop-agent = "modular.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["modular*"]

[tool.black]
line-length = 88
target-version = ['py38']
